back/pyproject.toml
[project]
name = "Chat-Destino"
version = "0.1.0"
description = "Chat de conversa para a destino"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.83",
    "langchain-community>=0.3.83",
    "langchain-core>=0.3.83",
    "langchain-chroma>=0.1.4",
    "langchain-text-splitters>=0.3.4",
    "pypdf>=6.9.2",
    "langchain-ollama>=0.3.10",
    "fastapi[standard]>=0.135.3",
    "websockets>=16.0",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
    "asyncpg>=0.31.0",
]

[tool.fastapi]
entrypoint = "src.main:app"
